small fix in CloudCoinFile.cs

+++ CloudCoin SafeScan/UI/HowMuchWindow.xaml.cs
@@ -41,7 +41,7 @@ namespace CloudCoin_SafeScan
 
         private void onOKClicked(object sender, RoutedEventArgs e)
         {
-            short sum = short.Parse(enterSumBox.Text);
+            long sum = long.Parse(enterSumBox.Text);
             if (sum <= Safe.Instance?.Contents.SumInStack)
             {
                 DialogResult = true;
